+++ CHANGELOG.md
@@ -12,6 +12,7 @@ From version 2.6.0, the sections in this file adhere to the [keep a changelog](h
 * [#2112](https://github.com/Shopify/shopify-cli/pull/2112): Fix intermittent error ("can't add a new key into hash during iteration") in the `theme push` command
 * [#2088](https://github.com/Shopify/shopify-cli/pull/2088): Update theme-check to 1.10.1
 * [#2130](https://github.com/Shopify/shopify-cli/pull/2130): Fix Homebrew installation.
+* [#2133](https://github.com/Shopify/shopify-cli/pull/2133): Fix ignore file handling in DevServer::Watcher.
 
 ## Version 2.13.0